# gh-92216: improve performance of hasattr for type objects