--- a/Misc/NEWS.d/next/Core and Builtins/2022-12-04-00-38-33.gh-issue-92216.CJXuWB.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-12-04-00-38-33.gh-issue-92216.CJXuWB.rst
@@ -1,0 +1,1 @@
+Improve the performance of :func:`hasattr` for type objects with a missing attribute.