This PR improves performance of `hasattr` for type objects in the case the attribute is missing. 

The optimization is similar to the one used for objects which have `PyObject_GenericGetAttr` as the attribute lookup method
(see [specialization for PyObject_GenericGetAttr in _PyObject_LookupAttr](https://github.com/python/cpython/blob/364ed9409269fb321dc4eafdea677c09a4bc0d8d/Objects/object.c#L952) and [suppress in _PyObject_GenericGetAttrWithDict](https://github.com/python/cpython/blob/364ed9409269fb321dc4eafdea677c09a4bc0d8d/Objects/object.c#L1348) )

Benchmark:
```
import pyperf
runner = pyperf.Runner()

setup="""
from dataclasses import dataclass, asdict
@dataclass
class D:
	a : dict
	b: list
	
d=D({'1': 1}, [1,2,3])
asdict(d)
"""

runner.timeit(name=f"micro benchmark", stmt=f"hasattr(o, '__str__')", setup='import os; o=type(os)')
runner.timeit(name=f"micro benchmark: misssing attr", stmt=f"hasattr(o, '_no_such_attr')", setup='import os; o=type(os)')
runner.timeit(name=f"asdict on dataclass", stmt=f"asdict(d)", setup=setup)
```
Results:
```
micro benchmark: Mean +- std dev: [main] 42.1 ns +- 1.1 ns -> [pr] 44.1 ns +- 1.8 ns: 1.05x slower
micro benchmark: misssing attr: Mean +- std dev: [main] 256 ns +- 6 ns -> [pr] 37.6 ns +- 0.7 ns: 6.80x faster
asdict on dataclass: Mean +- std dev: [main] 8.87 us +- 0.17 us -> [pr] 6.61 us +- 0.05 us: 1.34x faster

Geometric mean: 2.06x faster
```

<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
gh-NNNNN: Summary of the changes made
```

Where: gh-NNNNN refers to the GitHub issue number.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->


<!-- gh-issue-number: gh-92216 -->
* Issue: gh-92216
<!-- /gh-issue-number -->