@Fidget-Spinner 2022-12-05
Thanks! If I understand correctly, the speedup comes from not having to raise and clear an exception right?