@Fidget-Spinner 2022-12-22
@eendebakpt alright. Sorry I've been putting this off because the code it touches isn't something I'm super familiar with.

However I will take a look at the other examples you pointed out that uses a similar concept.